+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.3.0] - 2024-04-16
+
+### Added
+
+- added `Null`-type (`1657715`)
+
+### Fixed
+
+- fixed potentially problematic `Property` where the `name`-property is an empty string (`409551d`)
+- fixed issue where an `Object` with `model` causes crash when field is missing in input (`f7e30c2`)
+
 ## [0.2.0] - 2024-04-16
 
 ### Changed